Решена проблема выделения по одной ячейке в секции коллекции выбора эмоджи/цвета, данные названий секций, эмоджи и цветов вынесены в отдельную структуру
