+++ Tracker/Helpers/HelperData.swift
@@ -1,0 +1,25 @@
+//
+//  HelperData.swift
+//  Tracker
+//
+//  Created by Кирилл Арефьев on 23.05.2024.
+//
+
+import UIKit
+
+struct HelperData {
+    // MARK: - Static Properties
+    static let sectionTitles: [String] = ["Emoji", "Цвет"]
+    
+    static let emojis: [String] = [
+        "🙂", "😻", "🌺", "🐶", "❤️", "😱",
+        "😇", "😡", "🥶", "🤔", "🙌", "🍔",
+        "🥦", "🏓", "🥇", "🎸", "🏝", "😪"
+    ]
+    
+    static let selectionColors: [UIColor] = [
+        .color1, .color2, .color3, .color4, .color5, .color6,
+        .color7, .color8, .color9, .color10, .color11, .color12,
+        .color13, .color14, .color15, .color16, .color17, .color18
+    ]
+}
